src/main/java/com/nisum/evaluation/service/JwtTokenGenerator.java
package com.nisum.evaluation.service;

import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import java.util.Date;

@Component
public class JwtTokenGenerator {

    @Value("${nisum.app.token.generator-jwtSecret}")
    private String jwtSecret;

    @Value("${nisum.app.token.generator-jwtExpirationMs}")
    private int jwtExpirationMs;

    public String generateJwtToken(String userName) {
        return Jwts.builder()
                .setSubject(userName)
                .setIssuedAt(new Date())
                .setExpiration(new Date((new Date()).getTime() + jwtExpirationMs))
                .signWith(SignatureAlgorithm.HS512, jwtSecret)
                .compact();
    }
}
